Splitup the parser
