+++ src/utils/parser/utils.js
@@ -7,96 +7,10 @@ import { isDevelopment } from "devtools-config";
 import toPairs from "lodash/toPairs";
 import isEmpty from "lodash/isEmpty";
 import uniq from "lodash/uniq";
-import parseScriptTags from "parse-script-tags";
+import getAst, { traverseAst } from "./getAst";
+import { isFunction, isVariable } from "./shared";
 
 import type { SourceText, Location, Frame, TokenResolution } from "../../types";
-
-const ASTs = new Map();
-
-const symbolDeclarations = new Map();
-
-export type ASTLocation = {
-  start: {
-    line: number,
-    column: number
-  },
-  end: {
-    line: number,
-    column: number
-  }
-};
-
-export type SymbolDeclaration = {
-  name: string,
-  location: ASTLocation,
-  parameterNames?: string[]
-};
-
-export type SymbolDeclarations = {
-  functions: Array<SymbolDeclaration>,
-  variables: Array<SymbolDeclaration>
-};
-
-type Scope = {
-  location: {
-    line: number,
-    column: number
-  },
-  parent: Scope,
-  bindings: Object[]
-};
-
-function _parse(code, opts) {
-  return babylon.parse(
-    code,
-    Object.assign({}, opts, {
-      sourceType: "module",
-      plugins: ["jsx", "flow"]
-    })
-  );
-}
-
-function parse(text: string, opts?: Object) {
-  let ast;
-  if (!text) {
-    return;
-  }
-
-  try {
-    ast = _parse(text, opts);
-  } catch (error) {
-    if (isDevelopment()) {
-      console.warn("parse failed", text);
-    }
-
-    ast = {};
-  }
-
-  return ast;
-}
-
-function getAst(sourceText: SourceText) {
-  if (ASTs.has(sourceText.id)) {
-    return ASTs.get(sourceText.id);
-  }
-
-  let ast = {};
-  if (sourceText.contentType == "text/html") {
-    // Custom parser for parse-script-tags that adapts its input structure to
-    // our parser's signature
-    const parser = ({ source, line }) => {
-      return parse(source, {
-        startLine: line
-      });
-    };
-    ast = parseScriptTags(sourceText.text, parser) || {};
-  } else if (sourceText.contentType == "text/javascript") {
-    ast = parse(sourceText.text);
-  }
-
-  ASTs.set(sourceText.id, ast);
-  return ast;
-}
 
 function getNodeValue(node) {
   if (t.isThisExpression(node)) {
@@ -104,75 +18,6 @@ function getNodeValue(node) {
   }
 
   return node.name;
-}
-
-function getFunctionName(path): string {
-  if (path.node.id) {
-    return path.node.id.name;
-  }
-
-  const parent = path.parent;
-  if (parent.type == "ObjectProperty") {
-    return parent.key.name;
-  }
-
-  if (parent.type == "ObjectExpression" || path.node.type == "ClassMethod") {
-    return path.node.key.name;
-  }
-
-  if (parent.type == "VariableDeclarator") {
-    return parent.id.name;
-  }
-
-  if (parent.type == "AssignmentExpression") {
-    if (parent.left.type == "MemberExpression") {
-      return parent.left.property.name;
-    }
-
-    return parent.left.name;
-  }
-
-  return "anonymous";
-}
-
-function isFunction(path) {
-  return (
-    t.isFunction(path) ||
-    t.isArrowFunctionExpression(path) ||
-    t.isObjectMethod(path) ||
-    t.isClassMethod(path)
-  );
-}
-
-function getVariableNames(path): SymbolDeclaration[] {
-  if (t.isObjectProperty(path) && !isFunction(path.node.value)) {
-    return [
-      {
-        name: path.node.key.name,
-        location: path.node.loc
-      }
-    ];
-  }
-
-  if (!path.node.declarations) {
-    return path.node.params.map(dec => ({
-      name: dec.name,
-      location: dec.loc
-    }));
-  }
-
-  return path.node.declarations.map(dec => ({
-    name: dec.id.name,
-    location: dec.loc
-  }));
-}
-
-function isVariable(path) {
-  return (
-    t.isVariableDeclaration(path) ||
-    (isFunction(path) && path.node.params.length) ||
-    (t.isObjectProperty(path) && !isFunction(path.node.value))
-  );
 }
 
 function getMemberExpression(root) {
@@ -242,56 +87,9 @@ function isLexicalScope(path) {
   return t.isBlockStatement(path) || isFunction(path) || t.isProgram(path);
 }
 
-export function getSymbols(source: SourceText): SymbolDeclarations {
-  if (symbolDeclarations.has(source.id)) {
-    const symbols = symbolDeclarations.get(source.id);
-    if (symbols) {
-      return symbols;
-    }
-  }
-
-  const ast = getAst(source);
-
-  const symbols = { functions: [], variables: [] };
-
-  if (isEmpty(ast)) {
-    return symbols;
-  }
-
-  traverse(ast, {
-    enter(path) {
-      if (isVariable(path)) {
-        symbols.variables.push(...getVariableNames(path));
-      }
-
-      if (isFunction(path)) {
-        symbols.functions.push({
-          name: getFunctionName(path),
-          location: path.node.loc
-        });
-      }
-
-      if (t.isClassDeclaration(path)) {
-        symbols.variables.push({
-          name: path.node.id.name,
-          location: path.node.loc
-        });
-      }
-    }
-  });
-
-  symbolDeclarations.set(source.id, symbols);
-  return symbols;
-}
-
 function getClosestMemberExpression(source, token, location) {
-  const ast = getAst(source);
-  if (isEmpty(ast)) {
-    return null;
-  }
-
   let expression = null;
-  traverse(ast, {
+  traverseAst(source, {
     enter(path) {
       const { node } = path;
       if (
