+++ src/utils/parser/worker.js
@@ -1,4 +1,5 @@
-import { getSymbols, getVariablesInScope, resolveToken } from "./utils";
+import { getVariablesInScope, resolveToken } from "./utils";
+import getSymbols from "./getSymbols";
 
 import { workerUtils } from "devtools-utils";
 const { workerHandler } = workerUtils;
